Fix double free bug in Loader::close()
`_xsi_value_buffer` itself is a vector of pointers, whose entries are not popped.
Then during a second call to close() the vector size is unchanged and it still
contains pointers but to the previously deleted arrays. This causes the code to
enter the for loop again and try to delete[] the memory behind the pointers a
second time.

Fix this by explicitly emptying the vector.